COMP:Fixed warning C4267 on VS 64-bit.

+++ VTK/Graphics/vtkQuadraturePointStatistics.cxx
@@ -327,7 +327,7 @@ int vtkQuadraturePointStatistics::ComputeStatistics(
       }
     }
   results->GetFieldData()->Initialize();
-  return columns.size();
+  return static_cast<int>(columns.size());
 }
 
 //-----------------------------------------------------------------------------
